Fix heap_init flow -> trace
Previously, I'd get a false positive for trace on the ll test. This
occurred because after the possibility of a uaf was detected on the flow
test, the initialization of the trace used the final alias set of the
flow test, not the initial one, to initialize its own trace. This change
works around that by redoing the process of heap_init if a heap
dereference caused the issue.