Remove stack purging
It can cause incorrect behavior if the calling context points to
something freed in the function, but the function removes a pointer to
it. For example:

void f() {
  char* p = malloc(1);
  g(p);
  *p = 'a';
}
void g(char* p) {
  free(p);
  p = 0;
}

may end up forgetting that p was freed as it no longer knows that a
variable in the calling context could reference it.