Automate addition of mycroft files
Previously, new mycroft files needed a line in build.rs in addition to
their inclusion in datalog.rs to cause cargo to know when to rebuild for
them. Now, it build.rs will find all ".my" files in the mycroft
directory and add an implicit dependency on them to the crate.