Fixed (unrealized) bug in constraint ordering
Previously, all constraints were applied relative to the points-to set
on the instruction input.
In the event that a pointer was updated, then used, in separate BIL
statements but the same instruction, this interpretation would have
produced incorrect results.

Now, constraints are indexed first by bil statement order, then are an
unordered list of constraints applied by that statement. This allows
correct application both in the case where sequential instructions
modify the pointer then use it, and also for multiple-definitions
scenarios.