Use stack discipline to shrink pts
Previously, I tried this but ended up "forgetting" updates to values in
higher scopes if they were no longer reachable. This update instead
summarizes the upper scopes with a set of pointed to variables which
should be considered live even if they go out of scope in the inner
function.

This should get most of the optimization/readability of the previous
stack purging method without sacrificing correctness.